[CALCITE-3384] Support Kerberos-authentication using SPNEGO over HTTPS
Closes #113
