+++ core/src/main/java/org/apache/calcite/avatica/remote/AvaticaCommonsHttpClientImpl.java
@@ -102,7 +102,7 @@ public class AvaticaCommonsHttpClientImpl implements AvaticaHttpClient,
     initializeClient();
   }
 
-  private void initializeClient() {
+  protected void initializeClient() {
     socketFactoryRegistry = this.configureSocketFactories();
     configureConnectionPool(socketFactoryRegistry);
     this.authCache = new BasicAuthCache();
@@ -193,7 +193,7 @@ public class AvaticaCommonsHttpClientImpl implements AvaticaHttpClient,
     }
   }
 
-  public byte[] send(byte[] request) {
+  @Override public byte[] send(byte[] request) {
     while (true) {
       HttpClientContext context = HttpClientContext.create();
