+++ core/src/main/java/org/apache/calcite/avatica/remote/AvaticaCommonsHttpClientSpnegoImpl.java
@@ -16,7 +16,6 @@
  */
 package org.apache.calcite.avatica.remote;
 
-import org.apache.http.HttpHost;
 import org.apache.http.auth.AuthSchemeProvider;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.Credentials;
@@ -25,16 +24,13 @@ import org.apache.http.client.config.AuthSchemes;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.protocol.HttpClientContext;
-import org.apache.http.config.Lookup;
+import org.apache.http.config.Registry;
 import org.apache.http.config.RegistryBuilder;
+import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.apache.http.entity.ByteArrayEntity;
 import org.apache.http.entity.ContentType;
 import org.apache.http.impl.auth.SPNegoSchemeFactory;
-import org.apache.http.impl.client.BasicAuthCache;
 import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.apache.http.util.EntityUtils;
 
 import org.ietf.jgss.GSSCredential;
@@ -42,35 +38,23 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.net.URL;
 import java.security.Principal;
-import java.util.Objects;
-
 /**
  * Implementation of an AvaticaHttpClient which uses SPNEGO.
+ *
+ * ( at this point it could probably be just merged back into AvaticaCommonsHttpClientImpl)
  */
-public class AvaticaCommonsHttpClientSpnegoImpl implements AvaticaHttpClient {
+public class AvaticaCommonsHttpClientSpnegoImpl extends AvaticaCommonsHttpClientImpl {
   private static final Logger LOG = LoggerFactory
       .getLogger(AvaticaCommonsHttpClientSpnegoImpl.class);
 
   public static final String CACHED_CONNECTIONS_MAX_KEY = "avatica.http.spnego.max_cached";
-  public static final String CACHED_CONNECTIONS_MAX_DEFAULT = "100";
   public static final String CACHED_CONNECTIONS_MAX_PER_ROUTE_KEY =
       "avatica.http.spnego.max_per_route";
-  public static final String CACHED_CONNECTIONS_MAX_PER_ROUTE_DEFAULT = "25";
 
   private static final boolean USE_CANONICAL_HOSTNAME = true;
   private static final boolean STRIP_PORT_ON_SERVER_LOOKUP = true;
-
-  final URL url;
-  final HttpHost host;
-  final PoolingHttpClientConnectionManager pool;
-  final Lookup<AuthSchemeProvider> authRegistry;
-  final BasicCredentialsProvider credentialsProvider;
-  final BasicAuthCache authCache;
-  final CloseableHttpClient client;
 
   /**
    * Constructs an http client with the expectation that the user is already logged in with their
@@ -89,39 +73,39 @@ public class AvaticaCommonsHttpClientSpnegoImpl implements AvaticaHttpClient {
    * @param credential The GSS credentials
    */
   public AvaticaCommonsHttpClientSpnegoImpl(URL url, GSSCredential credential) {
-    this.url = Objects.requireNonNull(url);
+    super(url);
+    setGSSCredential(credential);
+  }
 
-    pool = new PoolingHttpClientConnectionManager();
-    // Increase max total connection to 100
+  @Override protected void configureConnectionPool(Registry<ConnectionSocketFactory> registry) {
+    super.configureConnectionPool(registry);
+    //For backwards compatibility, override the standard values if set
     final String maxCnxns =
-        System.getProperty(CACHED_CONNECTIONS_MAX_KEY, CACHED_CONNECTIONS_MAX_DEFAULT);
-    pool.setMaxTotal(Integer.parseInt(maxCnxns));
-    // Increase default max connection per route to 25
-    final String maxCnxnsPerRoute = System.getProperty(CACHED_CONNECTIONS_MAX_PER_ROUTE_KEY,
-        CACHED_CONNECTIONS_MAX_PER_ROUTE_DEFAULT);
-    pool.setDefaultMaxPerRoute(Integer.parseInt(maxCnxnsPerRoute));
+        System.getProperty(CACHED_CONNECTIONS_MAX_KEY);
+    if (maxCnxns != null) {
+      pool.setMaxTotal(Integer.parseInt(maxCnxns));
+    }
+    //For backwards compatibility, override the standard values if set
+    final String maxCnxnsPerRoute = System.getProperty(CACHED_CONNECTIONS_MAX_PER_ROUTE_KEY);
+    if (maxCnxnsPerRoute != null) {
+      pool.setDefaultMaxPerRoute(Integer.parseInt(maxCnxnsPerRoute));
+    }
+  }
 
-    this.host = new HttpHost(url.getHost(), url.getPort());
-
+  public void setGSSCredential(GSSCredential credential) {
     this.authRegistry = RegistryBuilder.<AuthSchemeProvider>create().register(AuthSchemes.SPNEGO,
         new SPNegoSchemeFactory(STRIP_PORT_ON_SERVER_LOOKUP, USE_CANONICAL_HOSTNAME)).build();
 
     this.credentialsProvider = new BasicCredentialsProvider();
     if (null != credential) {
       // Non-null credential should be used directly with KerberosCredentials.
+      // This is never set by the JDBC driver, nor the tests
       this.credentialsProvider.setCredentials(AuthScope.ANY, new KerberosCredentials(credential));
     } else {
       // A null credential implies that the user is logged in via JAAS using the
       // java.security.auth.login.config system property
       this.credentialsProvider.setCredentials(AuthScope.ANY, EmptyCredentials.INSTANCE);
     }
-
-    this.authCache = new BasicAuthCache();
-
-    // A single thread-safe HttpClient, pooling connections via the ConnectionManager
-    this.client = HttpClients.custom()
-        .setDefaultAuthSchemeRegistry(authRegistry)
-        .setConnectionManager(pool).build();
   }
 
   @Override public byte[] send(byte[] request) {
@@ -135,7 +119,7 @@ public class AvaticaCommonsHttpClientSpnegoImpl implements AvaticaHttpClient {
     ByteArrayEntity entity = new ByteArrayEntity(request, ContentType.APPLICATION_OCTET_STREAM);
 
     // Create the client with the AuthSchemeRegistry and manager
-    HttpPost post = new HttpPost(toURI(url));
+    HttpPost post = new HttpPost(uri);
     post.setEntity(entity);
 
     try (CloseableHttpResponse response = client.execute(post, context)) {
@@ -150,14 +134,6 @@ public class AvaticaCommonsHttpClientSpnegoImpl implements AvaticaHttpClient {
       throw e;
     } catch (Exception e) {
       LOG.debug("Failed to execute HTTP request", e);
-      throw new RuntimeException(e);
-    }
-  }
-
-  private static URI toURI(URL url) throws RuntimeException {
-    try {
-      return url.toURI();
-    } catch (URISyntaxException e) {
       throw new RuntimeException(e);
     }
   }
