+++ server/src/main/java/org/apache/calcite/avatica/server/HttpServer.java
@@ -779,9 +779,6 @@ public class HttpServer {
         handler = buildHandler(this, serverConfig);
         break;
       case SPNEGO:
-        if (usingTLS) {
-          throw new IllegalArgumentException("TLS has not been tested wtih SPNEGO");
-        }
         if (null != keytab) {
           LOG.debug("Performing Kerberos login with {} as {}", keytab, kerberosPrincipal);
           subject = loginViaKerberos(this);
